Create shim for release.sbom policy
The idea is that we could  create a shim (automaticaly) for every policy
we have. This way, policy rules could be included by either the old or
the new name. However, this may cause successes/violations/warnings to
appear twice in the report. This is not ideal. We could introduce a
change in EC CLI to handle deprecation or aliases.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>